+++ InterventionManagement.Business/Models/InterventionTemplate.cs
@@ -5,10 +5,10 @@ namespace au.edu.uts.ASDF.ENETCare.InterventionManagement.Business.Models
     public class InterventionTemplate
     {
         public string Name { get; set; }
-        public decimal Hours { get; set; }
-        public decimal Cost { get; set; }
+        public int Hours { get; set; }
+        public int Cost { get; set; }
 
-        public InterventionTemplate(string name, decimal hours, decimal cost)
+        public InterventionTemplate(string name, int hours, int cost)
         {
             if (!NameIsValid(name)) throw new InterventionTemplateNameException("The name was either empty, or missing.");
             if (!NumberIsPositive(hours)) throw new InterventionTemplateNumberException("Hours was not positive number");
@@ -26,7 +26,7 @@ namespace au.edu.uts.ASDF.ENETCare.InterventionManagement.Business.Models
             return !string.IsNullOrEmpty(name);
         }
 
-        private static bool NumberIsPositive(decimal number)
+        private static bool NumberIsPositive(int number)
         {
             return number >= 0;
         }
